Add all

+++ header/CPoint.h
@@ -1,10 +1,12 @@
 #ifndef __CPoint__
 #define __CPoint__
 
+#include <inttypes.h>
+
 
 struct SCoord{
-    unsigned int x;
-    unsigned int y;
+    uint16_t x;
+    uint16_t y;
 };
 
 class CPoint
@@ -13,14 +15,14 @@ private:
     SCoord coord;
     
 public:
-    CPoint(unsigned int x,unsigned int y);
+    CPoint(uint16_t x, uint16_t y);
     CPoint(SCoord c);
     ~CPoint();
-    float  getDistance(CPoint &p);
-    float  getDistance(SCoord &c);
-    SCoord getPoint();
-    bool   setPoint(SCoord &c);
-    bool   setPoint(CPoint &p);
+    float       getDistance(CPoint &p);
+    float       getDistance(SCoord &c);
+    SCoord      getPoint();
+    bool        setPoint(SCoord &c);
+    bool        setPoint(CPoint &p);
 };
 
 #endif /*__CPoint__*/
